[Bronze II] Title: Hashing, Time: 168 ms, Memory: 17764 KB, Score: 50 point - DashHub

+++ BOJ/Bronze/15829. Hashing/Main.java
@@ -9,7 +9,7 @@ class Main {
 		int r = 31;
 		int M = 1234567891;
 
-		int H = 0;
+		long H = 0;
 		for (int i = 0; i < L; i++) {
 			H += (s.charAt(i) - 'a' + 1) * Math.pow(r, i);
 		}
